requirements.txt
qrcode[pil]
django-extensions
Werkzeug
pyOpenSSL
xhtml2pdf
django-apscheduler
python-decouple